backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlmodel==0.0.31
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==4.3.0
pydantic-settings==2.12.0
pytest==7.4.3
httpx==0.27.0
pytest-asyncio==0.23.5
openai==1.57.4
mcp-sdk==1.0.0
python-multipart==0.0.20
alembic==1.13.1
asyncio==3.4.3